FINALLY get around to quieting the 64-to-32 bit integer warnings.

+++ XADMaster/Crypto/pbkdf2_hmac_sha256.c
@@ -6,17 +6,17 @@ void PBKDF2(const void *password,size_t passwordlength,
 const void *salt,size_t saltlength,
 uint8_t *DK,size_t DKlength,int count)
 {
-	int numblocks=(DKlength+HMAC_SHA256_DIGEST_LENGTH-1)/HMAC_SHA256_DIGEST_LENGTH;
+	int numblocks=(unsigned int)((DKlength+HMAC_SHA256_DIGEST_LENGTH-1)/HMAC_SHA256_DIGEST_LENGTH);
 	for(int i=0;i<numblocks;i++)
 	{
 		uint8_t Uj[HMAC_SHA256_DIGEST_LENGTH];
 
 		HMAC_SHA256_CTX ctx;
 		HMAC_SHA256_Init(&ctx);
-		HMAC_SHA256_UpdateKey(&ctx,password,passwordlength);
+		HMAC_SHA256_UpdateKey(&ctx,password,(unsigned int)passwordlength);
 		HMAC_SHA256_EndKey(&ctx);
 		HMAC_SHA256_StartMessage(&ctx);
-		HMAC_SHA256_UpdateMessage(&ctx,salt,saltlength);
+		HMAC_SHA256_UpdateMessage(&ctx,salt,(unsigned int)saltlength);
 		HMAC_SHA256_UpdateMessage(&ctx,(uint8_t[4]) {
 			((i+1)>>24)&0xff,
 			((i+1)>>16)&0xff,
@@ -33,7 +33,7 @@ uint8_t *DK,size_t DKlength,int count)
 		{
 			HMAC_SHA256_CTX ctx;
 			HMAC_SHA256_Init(&ctx);
-			HMAC_SHA256_UpdateKey(&ctx,password,passwordlength);
+			HMAC_SHA256_UpdateKey(&ctx,password,(unsigned int)passwordlength);
 			HMAC_SHA256_EndKey(&ctx);
 			HMAC_SHA256_StartMessage(&ctx);
 			HMAC_SHA256_UpdateMessage(&ctx,Uj,HMAC_SHA256_DIGEST_LENGTH);
@@ -54,17 +54,17 @@ const void *salt,size_t saltlength,
 uint8_t *DK1,uint8_t *DK2,uint8_t *DK3,size_t DKlength,
 int count1,int count2,int count3)
 {
-	int numblocks=(DKlength+HMAC_SHA256_DIGEST_LENGTH-1)/HMAC_SHA256_DIGEST_LENGTH;
+	int numblocks=(int)((DKlength+HMAC_SHA256_DIGEST_LENGTH-1)/HMAC_SHA256_DIGEST_LENGTH);
 	for(int i=0;i<numblocks;i++)
 	{
 		uint8_t Uj[HMAC_SHA256_DIGEST_LENGTH];
 
 		HMAC_SHA256_CTX ctx;
 		HMAC_SHA256_Init(&ctx);
-		HMAC_SHA256_UpdateKey(&ctx,password,passwordlength);
+		HMAC_SHA256_UpdateKey(&ctx,password,(unsigned int)passwordlength);
 		HMAC_SHA256_EndKey(&ctx);
 		HMAC_SHA256_StartMessage(&ctx);
-		HMAC_SHA256_UpdateMessage(&ctx,salt,saltlength);
+		HMAC_SHA256_UpdateMessage(&ctx,salt,(unsigned int)saltlength);
 		HMAC_SHA256_UpdateMessage(&ctx,(uint8_t[4]) {
 			((i+1)>>24)&0xff,
 			((i+1)>>16)&0xff,
@@ -81,7 +81,7 @@ int count1,int count2,int count3)
 		{
 			HMAC_SHA256_CTX ctx;
 			HMAC_SHA256_Init(&ctx);
-			HMAC_SHA256_UpdateKey(&ctx,password,passwordlength);
+			HMAC_SHA256_UpdateKey(&ctx,password,(unsigned int)passwordlength);
 			HMAC_SHA256_EndKey(&ctx);
 			HMAC_SHA256_StartMessage(&ctx);
 			HMAC_SHA256_UpdateMessage(&ctx,Uj,HMAC_SHA256_DIGEST_LENGTH);
@@ -99,7 +99,7 @@ int count1,int count2,int count3)
 		{
 			HMAC_SHA256_CTX ctx;
 			HMAC_SHA256_Init(&ctx);
-			HMAC_SHA256_UpdateKey(&ctx,password,passwordlength);
+			HMAC_SHA256_UpdateKey(&ctx,password,(unsigned int)passwordlength);
 			HMAC_SHA256_EndKey(&ctx);
 			HMAC_SHA256_StartMessage(&ctx);
 			HMAC_SHA256_UpdateMessage(&ctx,Uj,HMAC_SHA256_DIGEST_LENGTH);
@@ -115,7 +115,7 @@ int count1,int count2,int count3)
 		{
 			HMAC_SHA256_CTX ctx;
 			HMAC_SHA256_Init(&ctx);
-			HMAC_SHA256_UpdateKey(&ctx,password,passwordlength);
+			HMAC_SHA256_UpdateKey(&ctx,password,(unsigned int)passwordlength);
 			HMAC_SHA256_EndKey(&ctx);
 			HMAC_SHA256_StartMessage(&ctx);
 			HMAC_SHA256_UpdateMessage(&ctx,Uj,HMAC_SHA256_DIGEST_LENGTH);
