+++ XADMaster/XADNDSParser.m
@@ -12,7 +12,7 @@ static NSData *ConvertTiledIconToPNG(uint8_t *tiledata,uint16_t *palette);
 +(BOOL)recognizeFileWithHandle:(CSHandle *)handle firstBytes:(NSData *)data name:(NSString *)name
 {
 	const uint8_t *bytes=[data bytes];
-	int length=[data length];
+	NSInteger length=[data length];
 
 	if(length<0x200) return NO;
 	if(bytes[0x12]!=0) return NO;
